move util.js to util/util.js
