+++ js/layers/network-layers.util.js
@@ -1,4 +1,4 @@
-import { fetchJSON } from "../util.js";
+import { fetchJSON } from "../util/util.js";
 
 export const loadNetworkLayer = async (endpoint, name) => {
 	const api_endpoint =
